+++ bpf/insn/samples/bpf_insn_samples.h
@@ -6,3 +6,5 @@
 
 struct bpf_insn *trace_printk_insns(size_t *cnt);
 struct bpf_insn *strncmp_insns(size_t *cnt);
+
+struct bpf_insn *cgroup_from_id_insns(size_t *cnt);